🐛 fix: UserRepository 수정

+++ mr_patent_backend/src/main/java/com/d208/mr_patent_backend/domain/user/repository/UserRepository.java
@@ -15,5 +15,5 @@ public interface UserRepository extends JpaRepository<User, Integer> {
 
     Optional<User> findByUserRefreshToken(String refreshToken);
 
-    Optional<User> findById(Long userId);
+    Optional<User> findById(Integer userId);
 }
